+++ R/run_corpus_explorer.R
@@ -22,6 +22,12 @@
 #'   not allow patterns that will result in an enormous amount of hits or will
 #'   lead to a very slow search. (Examples of such patterns will include
 #'   '\code{.}' and '\code{\\b}'.)
+#' @param ui_options List. Specify custom app settings. Currently available:
+#' \itemize{
+#'     \item \code{MAX_DOCS_IN_WALL_VIEW} specifies the maximum number of
+#'     documents to be rendered in the 'document wall' view.
+#'     Default value is 12000.
+#'     }
 #' @param ... Other arguments passed to \code{\link[shiny]{runApp}} in the Shiny
 #'   package.
 #' @export
@@ -37,11 +43,13 @@
 #' titles <- paste("Text", 1:10)
 #' test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)
 #'
-#' # Converting to corporaexplorer object:
+#' # Converting to corporaexplorerobject:
 #' corpus <- prepare_data(test_df, corpus_name = "Test corpus")
+#'
 #' if(interactive()){
 #' # Running exploration app:
-#' run_corpus_explorer(corpus)
+#' run_corpus_explorer(corpus,
+#'                     ui_options = list(MAX_DOCS_IN_WALL_VIEW = 12001))
 #'
 #' # Running app to extract documents:
 #' run_document_extractor(corpus)
@@ -55,6 +63,7 @@ run_corpus_explorer <- function(corpus_object,
                                 ),
                                 optional_info = FALSE,
                                 allow_unreasonable_patterns = FALSE,
+                                ui_options = list(),
                                 ...) {
   app <- system.file("explorer", "app.R", package = "corporaexplorer")
   if (app == "") {
@@ -111,6 +120,9 @@ run_corpus_explorer <- function(corpus_object,
   shiny::shinyOptions("corporaexplorer_regex_engine" = regex_engine)
   shiny::shinyOptions("corporaexplorer_use_matrix" = use_matrix)
   shiny::shinyOptions("corporaexplorer_allow_unreasonable_patterns" = allow_unreasonable_patterns)
+
+  shiny::shinyOptions("ui_options" = ui_options)
+
 
   data <- as.character(substitute(corpus_object))
   shiny::shinyOptions("corporaexplorer_data" = data)