+++ NEWS.md
@@ -1,5 +1,9 @@
 # corporaexplorer 0.6.2.9000
 
+## API changes
+
+* Add `ui_options` argument to run_corpus_explorer().
+
 ## Bug fixes
 
 * Print method for corporaexplorerobject now works properly